refactor(auth): extract district handoff
